Move drawArrow to edge class, remove unused functions, rename curvedEdge to bezier
